+++ common/ayon_common/distribution/addon_info.py
