+++ common/ayon_common/resources/eye.png
