+++ common/ayon_common/resources/openpype_icon.png
