+++ common/ayon_common/resources/stylesheet.css
